+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Reduced load on LINSTOR when attaching volumes.
+- Reduced load on LINSTOR when detaching volumes.
 - Reduced load on LINSTOR when checking node stats.
 - Reduced load on LINSTOR when listing volumes with status.